Update ImapClient.cs
When calling this method on a gmail account with unlimited storage, the storage size that comes back to bigger that an int32 will allow, and an exception is throw.  By changing this to int64, the large number is accepted, and the exception is not thrown.

+++ ImapClient.cs
@@ -2212,8 +2212,8 @@ namespace S22.Imap {
 					if (m.Success) {
 						try {
 							MailboxQuota quota = new MailboxQuota(m.Groups[2].Value,
-								UInt32.Parse(m.Groups[3].Value),
-								UInt32.Parse(m.Groups[4].Value));
+								UInt64.Parse(m.Groups[3].Value),
+								UInt64.Parse(m.Groups[4].Value));
 							quotas.Add(quota);
 						} catch {
 							throw new BadServerResponseException(response);
